STM32F4: Crude implementation that works with RDP = 0
With the current state, the device enters SRAM after power cycling it, patches the reset vector and
successfully dumps the firmware. The only issue is, this only works with rdp = 0.

It seems that a reset from SRAM execution to Flash execution no longer riddens the read out protection.
I'm afraid this may just hint the vulnerability is no longer present on F4 devices (at least on my F401)....

Signed-off-by: CTXz <ctx.xda@gmail.com>