Target + Attack: Workaround/Fallback solution for devices that experience reset loops
This workaround should make the exploit work even on devices that suffer inexplicable resets.
It does so by maintaining the address incrementer/iterator as a global variable. This way, after
a reset occured, the stage 2 code will simply continue dumping where it last left off. To ensure
that USART garbage that's generated upon a reset is not interpreted as false positive, the start
magic is send prior the transmission of every byte. This comes with a speed penatly, as it
increases the time for the dump to complete.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>